Take Node's default locale on Linux, and round as V8 does on arm64
On Linux with no locale set, or C or POSIX, ICU's default locale is its
fallback en_US_POSIX, which V8 makes "en-US" for Intl's default
(Isolate::DefaultLocale); the runtime took ICU's as it was, and
Intl.DurationFormat("lkt") resolved to "en-US-u-va-posix".

On a Mac, Date.UTC answered 34448384 where Node answers 34447360: go-intl's
Date arithmetic let arm64 fuse a multiply and an add into one rounding.
go-intl v0.1.1 rounds each product first. The engine's own parseInt and
number literals in a radix, Number.prototype.toString in one, and
Math.hypot could fuse the same way, and now round too. TestNoFusedArithmetic
compiles the engine for arm64 and fails on any fused instruction, so that
an amd64 machine catches one.

+++ internal/lexer/lexer.go
@@ -750,7 +750,8 @@ func (l *Lexer) scanNumber(tok Token) (Token, error) {
 				}
 				var v float64
 				for i := 0; i < len(digits); i++ {
-					v = v*base + float64(digits[i]-'0')
+					// Rounded before the sum, which arm64 would fuse.
+					v = float64(v*base) + float64(digits[i]-'0')
 				}
 				tok.Num = v
 				tok.Value = tok.Raw
@@ -829,7 +830,7 @@ func (l *Lexer) scanRadix(tok Token, start, radix int, valid func(byte) bool) (T
 		if !valid(c) {
 			break
 		}
-		v = v*float64(radix) + float64(hexVal(c))
+		v = float64(v*float64(radix)) + float64(hexVal(c))
 		l.pos++
 	}
 	if l.pos == digitStart {
